ai_agents/AI-Voice-Agent/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dataclasses-json==0.6.7
deepgram-sdk==3.3.1
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.35.3
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
pygame==2.6.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
